refactor(postgres): Extract and compile generator
